+++ AGMPowerCLICommonFunctions.ps1
@@ -321,10 +321,6 @@ Function Get-AGMAPIData ([String]$filtervalue,[String]$keyword, [string]$search,
             {
                 Test-AGMJSON $RestError 
             }
-            elseif ($resp -contains "JWT has expired") {
-                Get-AGMErrorMessage  -messagetoprint  $resp
-                return
-            }
             else
             {
             if ( (!($resp.items)) -or ($itemoverride -eq $true) )
@@ -453,7 +449,13 @@ Function Test-AGMJSON()
         }
         if (!$validJson) 
         {
-            Write-Host "$args"
+            if ($isthisjson = "OpenID Connect token expired: JWT has expired") 
+            {
+                Get-AGMErrorMessage  -messagetoprint $isthisjson 
+            }
+            else {
+                Get-AGMErrorMessage  -messagetoprint $isthisjson 
+            }
         }
         else
         {
@@ -462,12 +464,14 @@ Function Test-AGMJSON()
             if ($testoutput.err_message)
             {
                 $testoutput.err_message = $testoutput.err_message -replace "`n",","
+                
             }
-            if ($testoutput.err_code -eq 10011)
+            elseif ($testoutput.err_code -eq 10011)
             {
                 Get-AGMErrorMessage -messagetoprint "User does not have permission to perform this action" 
+                
             }
-            if ($testoutput.error)
+            elseif ($testoutput.error)
             {
                 $testoutput.error
             }