+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 Started Change log file as per release 0.0.0.39
 
+## AGMPowerCLI 0.0.0.45
+* Get-AGMSLTPolicy will now support -sltid as well as -id, plus -policyid to display just one policy and -settableoption to display any policy options
+* If the OpenID Connect token has expired, this will now print as an obvious errormessage
+
 ## AGMPowerCLI 0.0.0.44
 * If service account token generator role was missing, no error was printing in PS7.